ml-models/requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
